Refactor geo functions in their own file
