+++ packages/react-core/src/filters/tileFeatures.d.ts
@@ -1,6 +1,3 @@
-import { TileFeatures, TileFeaturesResponse, Viewport } from '../types';
-import { Polygon, MultiPolygon } from 'geojson';
+import { TileFeatures, TileFeaturesResponse } from '../types';
 
-export function getGeometryToIntersect(viewport: number[] | null, geometry: Polygon | MultiPolygon | null): Polygon | MultiPolygon | null;
-export function isGlobalViewport(viewport: Viewport | null): boolean;
 export function tileFeatures(arg: TileFeatures): TileFeaturesResponse;
